+++ src/key_value_store.js
@@ -11,6 +11,7 @@ import { checkParamOrThrow, parseBody } from 'apify-client/build/utils';
 import {
     addCharsetToContentType, apifyClient, ensureDirExists, openRemoteStorage, openLocalStorage, ensureTokenOrLocalStorageEnvExists,
 } from './utils';
+import { APIFY_API_BASE_URL } from './constants';
 
 export const LOCAL_STORAGE_SUBDIR = LOCAL_STORAGE_SUBDIRS.keyValueStores;
 const MAX_OPENED_STORES = 1000;
@@ -293,6 +294,17 @@ export class KeyValueStore {
     }
 
     /**
+     * Returns a URL for the given key that may be used to publicly
+     * access the value in the remote key value store.
+     *
+     * @param {string} key
+     * @return {string}
+     */
+    getPublicUrl(key) {
+        return `${APIFY_API_BASE_URL}/key-value-stores/${this.storeId}/records/${key}`;
+    }
+
+    /**
      * Iterates over key value store keys, yielding each in turn to an `iteratee` function.
      * Each invocation of `iteratee` is called with three arguments: `(item, index, info)`.
      *
@@ -511,6 +523,21 @@ export class KeyValueStoreLocal {
      */
     _getPath(fileName) {
         return path.resolve(this.localStoragePath, fileName);
+    }
+
+    /**
+     * Returns a file:// URL for the given fileName that may be used to
+     * access the value on the local drive.
+     *
+     * Unlike in the remote store where key is sufficient, a full fileName
+     * must be provided here including the extension for the URL to be valid.
+     *
+     * @param {string} fileName
+     * @return {string}
+     * @ignore
+     */
+    getPublicUrl(fileName) {
+        return `file://${this._getPath(fileName)}`;
     }
 }