+++ test/key_value_store.js
@@ -368,6 +368,25 @@ describe('KeyValueStore', () => {
         });
     });
 
+
+    describe('getPublicUrl', () => {
+        it('should return the local url of a file', () => {
+            process.env[ENV_VARS.LOCAL_STORAGE_DIR] = LOCAL_STORAGE_DIR;
+            const store = new KeyValueStoreLocal('my-store-id', LOCAL_STORAGE_DIR);
+            expect(store.getPublicUrl('file.txt')).to.equal(`file://${store.localStoragePath}/file.txt`);
+            delete process.env[ENV_VARS.LOCAL_STORAGE_DIR];
+        });
+
+        it('should return the url of a file in apify cloud', async () => {
+            process.env[ENV_VARS.TOKEN] = 'xxx';
+            const publicUrl = 'https://api.apify.com/v2/key-value-stores';
+            const store = new KeyValueStore('my-store-id');
+
+            expect(store.getPublicUrl('file')).to.equal(`${publicUrl}/my-store-id/records/file`);
+            delete process.env[ENV_VARS.TOKEN];
+        });
+    });
+
     describe('forEachKey', () => {
         it('should work remotely', async () => {
             const storeId = 'some-id-1';